refactor: globalize monitor batch start route
