+++ backend/monitor/api/http/global_router.py
@@ -3,10 +3,11 @@
 import asyncio
 from typing import Annotated
 
-from fastapi import APIRouter, Depends, HTTPException, Query
+from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query, Request
 from pydantic import BaseModel, Field
 
 from backend.monitor.api.http.dependencies import get_current_user_id
+from backend.monitor.api.http.execution_target import resolve_monitor_evaluation_base_url
 from backend.monitor.infrastructure.web import gateway as monitor_gateway
 
 router = APIRouter()
@@ -33,6 +34,16 @@ def _or_404_or_503(fn, *args):
         raise HTTPException(status_code=404, detail=str(exc)) from exc
     except RuntimeError as exc:
         raise HTTPException(status_code=503, detail=str(exc)) from exc
+
+
+def _extract_bearer_token(request: Request) -> str:
+    auth_header = request.headers.get("Authorization", "")
+    if not auth_header.startswith("Bearer "):
+        raise HTTPException(status_code=401, detail="Missing or invalid Authorization header")
+    token = auth_header.removeprefix("Bearer ").strip()
+    if not token:
+        raise HTTPException(status_code=401, detail="Missing or invalid Authorization header")
+    return token
 
 
 async def _resource_io(fn, *args):
@@ -116,6 +127,26 @@ def evaluation_batch_create_action(
         sandbox=payload.sandbox,
         max_concurrent=payload.max_concurrent,
     )
+
+
+@router.post("/evaluation/batches/{batch_id}/start")
+def evaluation_batch_start_action(
+    batch_id: str,
+    request: Request,
+    background_tasks: BackgroundTasks,
+    _user_id: Annotated[str, Depends(get_current_user_id)],
+):
+    try:
+        return monitor_gateway.start_evaluation_batch(
+            batch_id=batch_id,
+            execution_base_url=resolve_monitor_evaluation_base_url(request),
+            token=_extract_bearer_token(request),
+            schedule_task=background_tasks.add_task,
+        )
+    except KeyError as exc:
+        raise HTTPException(status_code=404, detail=str(exc)) from exc
+    except ValueError as exc:
+        raise HTTPException(status_code=400, detail=str(exc)) from exc
 
 
 @router.get("/evaluation/scenarios")
